pyproject.toml
[project]
name = "django-keycloak-sso"
version = "0.1.3"
description = "Reusable Django app for SSO integration"
authors = [{name="Amir", email="amirhossein6168@gmail.com"}]
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "Django>=3.2",
    "djangorestframework>=3",
    "requests>=2",
    "python-jose>=3",
    "drf-yasg>=1.21",
    "drf-spectacular>=0.27",
    "PyJWT>=2.8.0",
]

[project.urls]
#Homepage = "https://your-git-host.com/django-sso-package"

[build-system]
requires = [
    "setuptools>=61"
]
build-backend = "setuptools.build_meta"
